Moved cards into another package. Corrected GeoTopicPage.
