+++ src/main/java/pages/GroupsPage.java
@@ -1,10 +1,11 @@
 package pages;
 
+import cards.NewGroupCard;
 import com.codeborne.selenide.Selenide;
 import org.openqa.selenium.By;
 
 import static org.openqa.selenium.By.xpath;
-import static pages.NewGroupCard.CREATE_GROUP_CARD;
+import static cards.NewGroupCard.CREATE_GROUP_CARD;
 
 public class GroupsPage extends HelperBase {
     private static final By CREATE_GROUP_BUTTON = xpath(".//*[@class = 'create-group']");
