+++ src/test/java/tests/GroupTests.java
@@ -6,7 +6,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import pages.*;
-import pages.MessageCard;
+import cards.PostCard;
+import cards.MessageCard;
 
 
 import java.io.IOException;
